+++ modules/School_Setup/Rollover.php
@@ -81,8 +81,11 @@ foreach ( (array) $tables as $table => $name )
 
 	$checked = ( $exists_RET[$table][1]['COUNT'] > 0 ) ? '' : ' checked';
 
+	// Fix SQL error foreign keys: force roll Schools
+	$readonly = ( $table === 'schools' && ! $exists_RET[$table][1]['COUNT'] ) ? ' readonly' : '';
+
 	$table_list .= '<tr><td><label><input type="checkbox" value="Y" name="tables[' . $table . ']"' .
-		$checked . '>&nbsp;' . $input_title . '</label>';
+		$checked . $readonly . '>&nbsp;' . $input_title . '</label>';
 
 	if ( $table === 'courses' )
 	{